requirements.txt
google-generativeai>=0.3.2
streamlit>=1.44.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
chromadb>=0.5.3
langchain>=0.2.16
langchain-chroma>=0.1.3
librosa>=0.10.2
pypdfium2>=4.30.0
pyyaml>=6.0.2
streamlit-mic-recorder>=0.0.8
transformers>=4.44.2
pydub>=0.25.1
python-multipart